BACKEND/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==22.0.0
mysqlclient==2.2.4
tzdata==2024.2